requirements.txt
inquirer == 2.7.0
javalang == 0.13.0
dockerfile-parse == 1.1.0
Flask
pandas
networkx
argparse
GitPython == 3.1.12

